refactor: extract model download path helpers
