+++ src/main/sources/standalone/templateInputAssets.ts
@@ -2,7 +2,7 @@ import fs from 'fs'
 import path from 'path'
 import * as settings from '../../settings'
 import { download } from '../../lib/download'
-import { stripQueryParams } from '../../lib/comfyDownloadManager'
+import { stripQueryParams } from '../../lib/downloadFilename'
 import { loadTemplateJson } from './templateModels'
 import { TEMPLATE_INPUT_BASE } from './curatedTemplates'
 import type { InstallationRecord } from '../../installations'
