+++ src/main/sources/standalone/templateModels.ts
@@ -2,7 +2,7 @@ import fs from 'fs'
 import path from 'path'
 import { getActiveVenvDir } from '../../lib/pythonEnv'
 import { findSitePackages } from './envPaths'
-import { ALLOWED_EXTENSIONS, stripQueryParams } from '../../lib/comfyDownloadManager'
+import { ALLOWED_EXTENSIONS, stripQueryParams } from '../../lib/downloadFilename'
 import { fetchJSON } from '../../lib/fetch'
 import { RAW_TEMPLATES_BASE } from './curatedTemplates'
 import type { InstallationRecord } from '../../installations'
